# Conflicts:
#   server/index.js

+++ server/controllers/profileController.js
@@ -1,6 +1,38 @@
 var fileHandler = require('../fileHandler.js');
 var Profile = require('../models/profile.js');
 
+const account_type = {
+	ADMIN: 'Admin',
+	ADOPT: 'Adopt',
+	USER: 'User'
+};
+
+exports.getUID = async (req, res) => {
+	console.log("sending uid");
+	try {
+		id = req.uid;
+		if (typeof id !== "undefined") {
+			res.status(200).json({
+				success: true,
+				uid: id,
+				message: "Success"
+			});
+		}
+		else {
+			res.status(500).json({
+				success: false,
+				message: "Failed to get UID"
+			});
+		}
+	}
+	catch (err) {
+		res.status(400).json({
+			success: false,
+			message: 'Failed to get UID'
+		});
+	}
+}
+
 exports.submitProfile = async (req, res) => {
 	// console.log(req.body.bio);
 	// console.log(req.file);
@@ -13,7 +45,9 @@ exports.submitProfile = async (req, res) => {
 		userId: req.decoded.id,
 		bio: req.body.bio,
 		profile_img:  req.file,
-		pet: null
+		pet: null,
+		
+		addedlist: [Number],
 	});
 
 	try {
@@ -68,49 +102,52 @@ exports.getProfileById = async (req, res) => {
 }
 
 exports.updateProfile = async (req, res) => {
-	try {
-		if(Object.keys(req.files).length === 0) {
-			req.files = null;
-		} else {
-			console.log("image exists")
-			//console.log(req.files.pet_img[0]);
-			for(var key in req.files) {
-				req.files[key] = req.files[key][0];
-				let img_dir = req.files[key].destination + req.files[key].filename; 
-				req.files[key].buff = fileHandler.img2Base64(img_dir);
-				fileHandler.deleteFile(img_dir);
-				//console.log(req.files[key]);
+	let id = req.params.id;
+	if (req.account_type === account_type.ADMIN || req.uid == id) {
+		try {
+			if(Object.keys(req.files).length === 0) {
+				req.files = null;
+			} else {
+				console.log("image exists")
+				//console.log(req.files.pet_img[0]);
+				for(var key in req.files) {
+					req.files[key] = req.files[key][0];
+					let img_dir = req.files[key].destination + req.files[key].filename; 
+					req.files[key].buff = fileHandler.img2Base64(img_dir);
+					fileHandler.deleteFile(img_dir);
+					//console.log(req.files[key]);
+				}
 			}
-		}
-		console.log(req.body);
-		let id = req.params.id;
-		let update_fields = {
-								bio: req.body.bio, 
-								pet_name: req.body.pet_name,
-								pet_age: req.body.pet_age,
-								pet_description: req.body.pet_description,
-								files: req.files
-							};
-		let status = await Profile.updateProfile(id, update_fields);
+			console.log(req.body);
+			let id = req.params.id;
+			let update_fields = {
+									bio: req.body.bio, 
+									pet_name: req.body.pet_name,
+									pet_age: req.body.pet_age,
+									pet_description: req.body.pet_description,
+									files: req.files
+								};
+			let status = await Profile.updateProfile(id, update_fields);
 
-		console.log(status);
-		if(status.success) {
+			console.log(status);
+			if(status.success) {
+					res.status(200).json({
+							success: true,
+							message: 'Profile updated',
+					});
+			} else {
 				res.status(200).json({
-						success: true,
-						message: 'Profile updated',
+					success: false,
+					message: 'Profile found',
 				});
-		} else {
-			res.status(200).json({
-				success: false,
-				message: 'Profile found',
+			}
+		} catch (err) {
+			console.log("Error at profileController.updateProfile");
+			console.log(err);
+			res.status(500).json({
+				error:err,
+				message: "Profile updated failed"
 			});
 		}
-	} catch (err) {
-		console.log("Error at profileController.updateProfile");
-		console.log(err);
-		res.status(500).json({
-			error:err,
-			message: "Profile updated failed"
-		});
 	}
 }
